Conflicts:
	Lilla spelprojektet/Code/GeometryManager.cpp

+++ Lilla spelprojektet/Code/EnemyHandler.cpp
@@ -46,7 +46,7 @@ int EnemyHandler::update(float dt)
 				vector<float> target = ai->findTarget(enemies.at(i)->getCurrentWaypoint(), 1);
 				float x = target[0];
 				float y = target[1];
-				vector<Waypoint> wp = ai->findPath(enemies.at(i)->getCurrentWaypoint1D(), x + y*10, 1);
+				vector<Waypoint> wp = ai->findPath(enemies.at(i)->getCurrentWaypoint1D(), (int)(x + y*10), 1);
 				if(wp.size() > 0)
 					enemies.at(i)->setPath(wp);
 			}
