+++ Lilla spelprojektet/Code/Input.cpp
@@ -35,10 +35,10 @@ float Input::mousePitch()
 	float dy = 0;
 	if((mState->btnState & MK_LBUTTON) == 0 )
 	{
-		dy = D3DXToRadian(0.25f * static_cast<float>(mState->yPos - this->prevMousePos.y));
+		dy = (float)D3DXToRadian(0.25f * static_cast<float>(mState->yPos - this->prevMousePos.y));
 	}
 
-	this->prevMousePos.y = mState->yPos;
+	this->prevMousePos.y = (float)mState->yPos;
 
 	return dy* 20;
 }
@@ -61,10 +61,10 @@ float Input::mouseRotateY()
 	float dx = 0; 
 	if((mState->btnState & MK_LBUTTON) == 0)
 	{
-		dx = D3DXToRadian(0.25f * static_cast<float>(mState->xPos - this->prevMousePos.x));
+		dx = (float)D3DXToRadian(0.25f * static_cast<float>(mState->xPos - this->prevMousePos.x));
 	}
 
-	this->prevMousePos.x = mState->xPos;
+	this->prevMousePos.x = (float)mState->xPos;
 
 	return dx * 20;
 }
